Revert many code-style changes: They’d only make it a pain to merge back.

+++ 360Controller/_60Controller.cpp
@@ -56,6 +56,7 @@ public:
 static UInt32 GetMaxPacketSize(IOUSBPipe *pipe)
 {
     const IOUSBEndpointDescriptor *ed = pipe->GetEndpointDescriptor();
+    
     if(ed==NULL)
         return 0;
     else
@@ -129,13 +130,18 @@ void Xbox360Peripheral::ChatPadTimerAction(IOTimerEventSource *sender)
 	{
 		case tsToggle:
 			SendToggle();
-			if (serialActive) {
-				if (!serialHeard) {
+			if (serialActive)
+			{
+				if (!serialHeard)
+				{
 					serialActive = false;
 					serialGot = 2;
 				}
-			} else {
-				if (serialHeard) {
+			}
+			else
+			{
+				if (serialHeard)
+				{
 					serialTimerState = tsReset1;
 					serialResetCount = 0;
 					nextTime = 40;
@@ -145,14 +151,20 @@ void Xbox360Peripheral::ChatPadTimerAction(IOTimerEventSource *sender)
 			
 		case tsMiniToggle:
 			SendToggle();
-			if (serialHeard) {
+			if (serialHeard)
+			{
 				serialTimerState = tsSet1;
 				nextTime = 40;
-			} else {
+			}
+			else
+			{
 				serialResetCount++;
-				if (serialResetCount > 3) {
+				if (serialResetCount > 3)
+				{
 					serialTimerState = tsToggle;
-				} else {
+				}
+				else
+				{
 					serialTimerState = tsReset1;
 					nextTime = 40;
 				}
@@ -194,7 +206,8 @@ void Xbox360Peripheral::ChatPadTimerAction(IOTimerEventSource *sender)
 	sender->setTimeoutMS(nextTime);	// Todo: Make it take into account function execution time?
 	serialHeard = false;
 	// Make it happen after the timer's set, for minimum impact
-	switch (serialGot) {
+	switch (serialGot)
+	{
 		case 1:
 			SerialConnect();
 			break;
@@ -480,9 +493,12 @@ bool Xbox360Peripheral::QueueSerialRead(void)
     complete.action = SerialReadCompleteInternal;
     complete.parameter = serialInBuffer;
     err = serialInPipe->Read(serialInBuffer, 0, 0, serialInBuffer->getLength(), &complete);
-    if (err == kIOReturnSuccess) {
+    if (err == kIOReturnSuccess)
+	{
 		return true;
-	} else {
+	}
+    else
+	{
         IOLog("read - failed to start for chatpad (0x%.8x)\n",err);
         return false;
     }
@@ -505,8 +521,7 @@ bool Xbox360Peripheral::QueueWrite(const void *bytes,UInt32 length)
     complete.action=WriteCompleteInternal;
     complete.parameter=outBuffer;
     err=outPipe->Write(outBuffer,0,0,length,&complete);
-    if (err == kIOReturnSuccess)
-        return true;
+    if(err==kIOReturnSuccess) return true;
     else {
         IOLog("send - failed to start (0x%.8x)\n",err);
         return false;
@@ -605,7 +620,7 @@ static inline XBox360_SShort getAbsolute(XBox360_SShort value)
 #else
 #error Unknown CPU byte order
 #endif
-    return (reverse < 0) ? ~reverse : reverse;
+    return (reverse<0)?~reverse:reverse;
 }
 
 // Adjusts the report for any settings speciified by the user
@@ -785,9 +800,9 @@ void Xbox360Peripheral::PadConnect(void)
 			OSString::withCString("IOKitDebug"),
         };
         const OSObject *objects[] = {
-            OSNumber::withNumber((unsigned)1, 32),
+            OSNumber::withNumber((unsigned long long)1, 32),
 			getProperty("IOCFPlugInTypes"),
-            OSNumber::withNumber((unsigned)65535, 32),
+            OSNumber::withNumber((unsigned long long)65535, 32),
         };
         OSDictionary *dictionary = OSDictionary::withObjects(objects, keys, sizeof(keys) / sizeof(keys[0]));
 		if (padHandler->init(dictionary))
@@ -825,7 +840,7 @@ void Xbox360Peripheral::SerialConnect(void)
             OSString::withCString(kIOSerialDeviceType),
         };
         const OSObject *objects[] = {
-            OSNumber::withNumber((unsigned)0, 32),
+            OSNumber::withNumber((unsigned long long)0, 32),
         };
         OSDictionary *dictionary = OSDictionary::withObjects(objects, keys, sizeof(keys) / sizeof(keys[0]), 0);
         if (serialHandler->init(dictionary))
